Remove extra indirection for writing old binary results
We had two sets of builders/visitors on top of the actual thing that did the writing. Remove these since they did nothing other than be confusing
